feat: temp solution for modal (+ form) to be usable on mobile devices
(or when the height of the modal would exceed the window height - e.g. when using a form with a lots of fields)
The webpage itself is usable on mobile devices so the modal+form should be too!
There were no designs for modal+form on mobile device, so I went freestyle a bit.
Not ideal solution, but for now, it will suffice, the form is usable.