fix: ensure the search icon in the input is not covered by text
I tried to fix this particular issue multiple times and this seems to be the only viable option so far - until we address global styling, which is root of a lot of evil.
The placeholder for search on the homepage is a long string, while the width of the input it resides in is relatively small (and fixed), this results in the `search icon` to be misaligned.
The huge numer of global styling does not make it easier to figue out the fix (since there's also responsive design in play).